print time

+++ cuda-workspace/TestingCudo/FastGridMain.c
@@ -163,7 +163,7 @@ void computeFull(StateList * initState,Path ** path, int N,int depth, int thread
 		double time_spent = (double) (end - begin) / CLOCKS_PER_SEC;
 		//printf("Time spent %lf\n", time_spent);
 		printf("Time spent %lf\n", time_spent);
-		int value = time_spent % 60;
+		int value = (int)time_spent % 60;
 		printf("seconds %d ", value);
 		if(time_spent > 60)
 		{
